+++ cli/src/movement_util.rs
@@ -29,6 +29,7 @@ use crate::cli_util::WorkspaceCommandHelper;
 use crate::cli_util::short_commit_hash;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
+use crate::command_error::user_error_with_hint;
 use crate::ui::Ui;
 
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -157,6 +158,21 @@ fn get_target_commit(
     args: &MovementArgsInternal,
 ) -> Result<Commit, CommandError> {
     let wc_revset = RevsetExpression::commit(working_commit_id.clone());
+
+    // If we're not editing, the working-copy shouldn't have any children
+    if !args.should_edit
+        && !workspace_command
+            .repo()
+            .view()
+            .heads()
+            .contains(working_commit_id)
+    {
+        return Err(user_error_with_hint(
+            "The working copy must not have any children",
+            "Create a new commit on top of this one or use `--edit`",
+        ));
+    }
+
     // If we're editing, start at the working-copy commit. Otherwise, start from
     // its direct parent(s).
     let start_revset = if args.should_edit {